--- apps/frontend/component/ui/Input.tsx
@@ -1,46 +1,0 @@
-import React, { forwardRef } from 'react';
-import { clsx } from 'clsx';
-import { twMerge } from 'tailwind-merge';
-
-interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {
-  label?: string;
-  error?: string;
-  helperText?: string;
-  containerClassName?: string;
-}
-
-const Input = forwardRef<HTMLInputElement, InputProps>(
-  ({ label, error, helperText, className, containerClassName, id, ...props }, ref) => {
-    const inputId = id || props.name;
-
-    return (
-      <div className={twMerge('space-y-2', containerClassName)}>
-        {label && (
-          <label htmlFor={inputId} className="block text-sm font-medium text-gray-700">
-            {label}
-          </label>
-        )}
-        <div className="relative">
-          <input
-            ref={ref}
-            id={inputId}
-            className={twMerge(
-              'block w-full rounded-md border-gray-300 shadow-sm focus:border-blue-500 focus:ring-blue-500 sm:text-sm p-2 border text-gray-900 placeholder:text-gray-400',
-              error && 'border-red-500 focus:border-red-500 focus:ring-red-500',
-              className
-            )}
-            {...props}
-          />
-        </div>
-        {helperText && !error && (
-          <p className="text-xs text-gray-500">{helperText}</p>
-        )}
-        {error && <p className="text-sm text-red-500">{error}</p>}
-      </div>
-    );
-  }
-);
-
-Input.displayName = 'Input';
-
-export default Input;
