+++ apps/frontend/components/dashboard/EscrowFilters.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import Input from '@/component/ui/Input';
+import Input from '@/components/ui/Input';
 
 interface EscrowFiltersProps {
   searchQuery: string;
