+++ apps/frontend/components/escrow/create/BasicInfoStep.tsx
@@ -4,7 +4,7 @@ import { useFormContext } from 'react-hook-form';
 import { CreateEscrowFormData } from '@/lib/escrow-schema';
 import Input from '../../ui/Input';
 import TextArea from '../../ui/TextArea';
-import Select from '../../ui/Select';
+import Select from '../../ui/BasicSelect';
 
 export default function BasicInfoStep() {
   const {
