+++ apps/frontend/components/homepage/Hero.tsx
