+++ apps/frontend/components/homepage/KeyFeatures.tsx
