+++ apps/frontend/components/ui/input.tsx
@@ -1,21 +1,46 @@
-import * as React from "react"
+import React, { forwardRef } from 'react';
+import { clsx } from 'clsx';
+import { twMerge } from 'tailwind-merge';
 
-import { cn } from "@/lib/utils"
-
-function Input({ className, type, ...props }: React.ComponentProps<"input">) {
-  return (
-    <input
-      type={type}
-      data-slot="input"
-      className={cn(
-        "file:text-foreground placeholder:text-muted-foreground selection:bg-primary selection:text-primary-foreground dark:bg-input/30 border-input h-9 w-full min-w-0 rounded-md border bg-transparent px-3 py-1 text-base shadow-xs transition-[color,box-shadow] outline-none file:inline-flex file:h-7 file:border-0 file:bg-transparent file:text-sm file:font-medium disabled:pointer-events-none disabled:cursor-not-allowed disabled:opacity-50 md:text-sm",
-        "focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:ring-[3px]",
-        "aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive",
-        className
-      )}
-      {...props}
-    />
-  )
+interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {
+  label?: string;
+  error?: string;
+  helperText?: string;
+  containerClassName?: string;
 }
 
-export { Input }
+const Input = forwardRef<HTMLInputElement, InputProps>(
+  ({ label, error, helperText, className, containerClassName, id, ...props }, ref) => {
+    const inputId = id || props.name;
+
+    return (
+      <div className={twMerge('space-y-2', containerClassName)}>
+        {label && (
+          <label htmlFor={inputId} className="block text-sm font-medium text-gray-700">
+            {label}
+          </label>
+        )}
+        <div className="relative">
+          <input
+            ref={ref}
+            id={inputId}
+            className={twMerge(
+              'block w-full rounded-md border-gray-300 shadow-sm focus:border-blue-500 focus:ring-blue-500 sm:text-sm p-2 border text-gray-900 placeholder:text-gray-400',
+              error && 'border-red-500 focus:border-red-500 focus:ring-red-500',
+              className
+            )}
+            {...props}
+          />
+        </div>
+        {helperText && !error && (
+          <p className="text-xs text-gray-500">{helperText}</p>
+        )}
+        {error && <p className="text-sm text-red-500">{error}</p>}
+      </div>
+    );
+  }
+);
+
+Input.displayName = 'Input';
+
+export default Input;
