+++ apps/frontend/components/ui/textarea.tsx
@@ -1,18 +1,46 @@
-import * as React from "react"
+import React, { forwardRef } from 'react';
+import { clsx } from 'clsx';
+import { twMerge } from 'tailwind-merge';
 
-import { cn } from "@/lib/utils"
-
-function Textarea({ className, ...props }: React.ComponentProps<"textarea">) {
-  return (
-    <textarea
-      data-slot="textarea"
-      className={cn(
-        "border-input placeholder:text-muted-foreground focus-visible:border-ring focus-visible:ring-ring/50 aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive dark:bg-input/30 flex field-sizing-content min-h-16 w-full rounded-md border bg-transparent px-3 py-2 text-base shadow-xs transition-[color,box-shadow] outline-none focus-visible:ring-[3px] disabled:cursor-not-allowed disabled:opacity-50 md:text-sm",
-        className
-      )}
-      {...props}
-    />
-  )
+interface TextAreaProps extends React.TextareaHTMLAttributes<HTMLTextAreaElement> {
+  label?: string;
+  error?: string;
+  helperText?: string;
+  containerClassName?: string;
 }
 
-export { Textarea }
+const TextArea = forwardRef<HTMLTextAreaElement, TextAreaProps>(
+  ({ label, error, helperText, className, containerClassName, id, ...props }, ref) => {
+    const inputId = id || props.name;
+
+    return (
+      <div className={twMerge('space-y-2', containerClassName)}>
+        {label && (
+          <label htmlFor={inputId} className="block text-sm font-medium text-gray-700">
+            {label}
+          </label>
+        )}
+        <div className="relative">
+          <textarea
+            ref={ref}
+            id={inputId}
+            className={twMerge(
+              'block w-full rounded-md border-gray-300 shadow-sm focus:border-blue-500 focus:ring-blue-500 sm:text-sm p-2 border text-gray-900 placeholder:text-gray-400',
+              error && 'border-red-500 focus:border-red-500 focus:ring-red-500',
+              className
+            )}
+            {...props}
+          />
+        </div>
+        {helperText && !error && (
+          <p className="text-xs text-gray-500">{helperText}</p>
+        )}
+        {error && <p className="text-sm text-red-500">{error}</p>}
+      </div>
+    );
+  }
+);
+
+TextArea.displayName = 'TextArea';
+
+export default TextArea;
